fixed test suites for mysql 5.5

+++ src/test/resources/sql/DeclarePrimaryKeysSuite.mysql.sql
@@ -9,7 +9,7 @@ create table Price (
 	currency varchar(3) not null,
 	unitprice decimal(6,3),
 	saleprice decimal(6,3),
-	product_id int not null,
+	product_id bigint unsigned not null,
 	primary key (product_id,currency,unitprice),
 	foreign key (product_id) references Product(id)
 )
