+++ src/test/resources/sql/ManyToManySimpleTypesSuite.mysql.sql
@@ -14,8 +14,8 @@ create table Category (
 ;
 
 create table Product_Category (
-	product_id int not null,
-	category_id int not null,
+	product_id bigint unsigned not null,
+	category_id bigint unsigned not null,
 	primary key (product_id,category_id),
 	foreign key (product_id) references Product(id) on delete cascade,
 	foreign key (category_id) references Category(id) on delete cascade
